Completed Game 1 instructions, removed unused game instruction files, updating routing in app.js and game 1 home page
