+++ frontend/src/App.js
@@ -11,7 +11,7 @@ import ReachAndRecallLevelsPage from "./components/game3/home_page";
 import ReachAndRecallMemorize from "./components/game3/memorize_numbers";
 import BalanceQuest from "./components/game1/game_logic";
 import FinalScore from './components/game3/final_score';
-import Game1Instructions from "./components/game1/game-1-instructions/game-1-instructions";
+import Game1Instructions from "./components/game1/game-1-instructions/Game1Instructions";
 import Game3Instructions from "./components/game3/game-3-instructions/Game3Instructions";
 import BalanceQuestLevelsPage from "./components/game1/home_page";
 import Settings from './components/settings';
@@ -52,7 +52,7 @@ const App = () => {
         <Route path="/reach-and-recall/:uid/final-score" element={<FinalScore user={user} />} />
         <Route path="/reach-and-recall/:uid/instructions/level/:level" element={<Game3Instructions user={user} />} />
         <Route path="/balance-quest/:uid/home-page" element={<BalanceQuestLevelsPage user={user} />} />
-        <Route path="/balance-quest/:uid/game/level/:level" element={<Game1Instructions />} />
+        <Route path="/balance-quest/:uid/game/level/:level" element={<BalanceQuest user={user} />} />
         <Route path="/balance-quest/:uid/instructions" element={<Game1Instructions user={user} />} />
       </Routes>
     </Router>
