--- frontend/src/components/game1/game-1-instructions/game-1-instructions.css
@@ -1,107 +1,0 @@
-.instructions-container {
-  text-align: center;
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: flex-start !important;
-  height: 100vh;
-  padding: 0 !important;
-  margin: 0;
-  overflow: hidden; /* Prevents scrolling */
-}
-
-.back-button {
-  position: absolute;
-  top: 15px;
-  left: 15px;
-  background: none;
-  border: none;
-  cursor: pointer;
-  transition: transform 0.2s ease-in-out;
-}
-
-.back-button:hover {
-  transform: scale(1.5);
-  background: none;
-  border: none;
-}
-
-.back-arrow {
-  width: 30px;
-  height: auto;
-}
-
-.game-title {
-  font-size: 25px;
-  font-weight: bold;
-  margin-top: 10px;
-  color: #B45522;
-}
-
-.instruction-subtext {
-  font-size: 40px;
-  font-weight: bold;
-  text-align: center;
-  margin-top: 5px;
-  color: #B45522;
-}
-
-/* Fruits Animation */
-.fruits-container {
-  display: flex;
-  justify-content: space-around;
-  align-items: flex-start;
-  gap: 50px;
-  margin-top: 50px;
-  position: relative;
-  height: 400px;
-  width: 100%;
-}
-
-.fruit {
-  width: 200px;
-  height: auto;
-  position: absolute;
-}
-
-.fruit-0 {
-  left: 50%;
-  transform: translateX(-50%);
-  animation: fall-straight 3s ease-in forwards;
-}
-
-.fruit-1 {
-  left: 20%;
-  animation: move-right-then-fall 3s ease-in forwards 3s;
-}
-
-.fruit-2 {
-  right: 20%;
-  transform: translateX(0);
-  animation: move-center-then-fall 3s ease-in forwards 6s;
-}
-
-@keyframes fall-straight {
-  0% { transform: translateX(-50%) translateY(0); opacity: 1; }
-  100% { transform: translateX(-50%) translateY(300px); opacity: 0; }
-}
-
-@keyframes move-right-then-fall {
-  0% { transform: translateX(0) translateY(0); opacity: 1; }
-  50% { transform: translateX(100%) translateY(0); opacity:1; }
-  100% { transform: translateX(100%) translateY(300px); opacity: 0; }
-}
-
-@keyframes move-center-then-fall {
-  0% { transform: translateY(0); opacity: 1; }
-  50% { transform: translateX(-100%) translateY(0); opacity:1;}
-  100% { transform: translateX(-100%) translateY(300px); opacity: 0; }
-}
-
-.bucket {
-  width: 500px;
-  height: auto;
-  position: absolute;
-  bottom: 0px;
-}
